major refactor: ThesisEditPage
